refactoring: Moves Chunk* classes under the .table package.
