+++ src/main/java/org/bricolages/streaming/preproc/Preprocessor.java
@@ -1,6 +1,7 @@
 package org.bricolages.streaming.preproc;
 import org.bricolages.streaming.event.*;
 import org.bricolages.streaming.stream.*;
+import org.bricolages.streaming.table.*;
 import org.bricolages.streaming.object.*;
 import org.bricolages.streaming.exception.*;
 import org.springframework.beans.factory.annotation.Autowired;
