+++ src/main/java/org/bricolages/streaming/table/ChunkProperties.java
@@ -1,0 +1,10 @@
+package org.bricolages.streaming.table;
+import java.time.Instant;
+
+public interface ChunkProperties {
+    int getObjectRows();
+    int getErrorRows();
+    String getObjectUrl();
+    long getObjectSize();
+    Instant getObjectCreatedTime();
+}
